requirement.txt
Django==5.2.1
djangorestframework==3.16.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
drf-spectacular==0.28.0